DDD-92 Add discounted price calculator strategy
