+++ src/main/java/lt/vu/application/order/service/PriceCalculator.java
@@ -5,21 +5,6 @@ import lt.vu.persistence.orm.entities.Order;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-@ApplicationScoped
-public class PriceCalculator {
-
-    @Inject
-    private Iterable<ChainablePriceCalculator> calculators;
-
-    public int calculate(Order order) {
-        int totalPrice = order.getPackageOption().getPrice();
-
-        for (ChainablePriceCalculator calculator : this.calculators) {
-            if (calculator.canCalculate(order)) {
-                totalPrice = calculator.calculate(totalPrice);
-            }
-        }
-
-        return totalPrice;
-    }
+public interface PriceCalculator {
+    int calculate(Order order);
 }
