+++ src/main/webapp/WEB-INF/beans.xml
@@ -9,4 +9,7 @@
     <interceptors>
         <class>lt.vu.infrastructure.interceptors.MethodLogger</class>
     </interceptors>
+    <alternatives>
+        <class>lt.vu.application.order.service.DiscountedPriceCalculator</class>
+    </alternatives>
 </beans>
